extract fixtures to separate file
